# src/types/common.types.ts
export interface BaseModel {
  id: string
  createdAt: Date
  updatedAt: Date
  deletedAt?: Date
}

export interface PaginationOptions {
  page?: number
  limit?: number
  sortBy?: string
  sortOrder?: "ASC" | "DESC"
}

export interface PaginationResult<T> {
  data: T[]
  pagination: {
    currentPage: number
    totalPages: number
    totalItems: number
    itemsPerPage: number
    hasNextPage: boolean
    hasPreviousPage: boolean
  }
}

export interface ApiResponse<T = any> {
  success: boolean
  message: string
  data?: T
  errors?: string[]
  timestamp: string
}

export interface UserContext {
  userId: string
  entityId: string
  fundIds: string[]
  roles: string[]
  permissions: string[]
}